libhashtable: fix UB in MAX_CAPACITY shift, drop now-wrong ssize_t cast
(1 << 31) shifts a 1 into the sign bit of a signed 32-bit int,
which is UB per C11 6.5.7p4 and trips UBSan on every executor
fuzz run within the first few thousand inputs.

Replace with ((size_t)1 << 31) so the shift happens in unsigned
type at the natural width and yields the intended value as size_t.

The capacity comparison at __ht_rehash() previously cast
ht->capacity (size_t) down to ssize_t to dance around the int
type of the old MAX_CAPACITY. With MAX_CAPACITY now properly
size_t, that cast is wrong-direction and introduces a sign-compare
warning. Drop it; the comparison is naturally size_t on both sides.

Issue #62.

+++ src/libhashtable/ht.c
@@ -21,8 +21,8 @@
 #define MAX_LOAD_FACTOR                                                        \
     (0.75) // Capacity point at which a table needs to grow and rehash
 #define MAX_CAPACITY                                                           \
-    (1 << 31) // Maximum capacity of table when it should not grow and rehash
-              // (2147483648)
+    ((size_t)1 << 31) // Maximum capacity of table when it should not grow and
+                      // rehash (2147483648)
 #define GROWTH_FACTOR (2) // Factor by which a table's capacity should grow
 
 typedef struct ht_bucket {
@@ -188,7 +188,7 @@ static void __ht_rehash(ht_t *ht) {
     size_t capacity;
 
     if (ht->used_buckets + 1 < (size_t)(ht->capacity * MAX_LOAD_FACTOR) ||
-        (ssize_t)ht->capacity >= MAX_CAPACITY) {
+        ht->capacity >= MAX_CAPACITY) {
         return;
     }
